fix: incorrect timestamps (#58)
# 🤖 Linear

Closes GIT-171

## Description
`block.timestamp` is treated as Unix timestamp in seconds, reason why
Envio client returns it as that. However, the system manages Unix
timestamp in milliseconds to be compliant with Javascript treatment in
Date objects (and it's also more usual to work it like this) and also
with external providers like Coingecko that also manages timestamps in
ms.

To fix this, we convert timestamps to MS in Envio indexer client. Also,
we create a branded type TimestampMs.


## Checklist before requesting a review

-   [x] I have conducted a self-review of my code.
-   [x] I have conducted a QA.
-   [x] If it is a core feature, I have included comprehensive tests.

+++ scripts/migrations/src/migrations/20241216T160549_event_registry.ts
@@ -13,7 +13,7 @@ export async function up(db: Kysely<any>): Promise<void> {
         .createTable("events_registry")
         .addColumn("chainId", CHAIN_ID_TYPE)
         .addColumn("blockNumber", "integer")
-        .addColumn("blockTimestamp", "integer")
+        .addColumn("blockTimestamp", "bigint")
         .addColumn("logIndex", "integer")
         .addColumn("rawEvent", "jsonb")
         .addPrimaryKeyConstraint("events_registry_pkey", ["chainId"])
